refactor: move `verify_filter` to new filter gadget
