+++ crates/proof-of-sql/src/sql/proof_gadgets/mod.rs
@@ -1,6 +1,7 @@
 //! This module contains shared proof logic for multiple `ProofExpr` / `ProofPlan` implementations.
 #[cfg(test)]
 mod divide_and_modulo_expr;
+pub(crate) mod filter_base;
 mod membership_check;
 mod monotonic;
 #[cfg_attr(not(test), expect(dead_code))]
