+++ crates/proof-of-sql/src/sql/proof_plans/filter_exec.rs
@@ -16,6 +16,7 @@ use crate::{
             ProverHonestyMarker, SumcheckSubpolynomialType, VerificationBuilder,
         },
         proof_exprs::{AliasedDynProofExpr, DynProofExpr, ProofExpr, TableExpr},
+        proof_gadgets::filter_base::verify_filter,
     },
     utils::log,
 };
@@ -280,49 +281,6 @@ impl ProverEvaluate for FilterExec {
 
         Ok(res)
     }
-}
-
-#[expect(clippy::similar_names)]
-pub(super) fn verify_filter<S: Scalar>(
-    builder: &mut impl VerificationBuilder<S>,
-    c_fold_eval: S,
-    d_fold_eval: S,
-    chi_n_eval: S,
-    chi_m_eval: S,
-    s_eval: S,
-) -> Result<(), ProofError> {
-    let c_star_eval = builder.try_consume_final_round_mle_evaluation()?;
-    let d_star_eval = builder.try_consume_final_round_mle_evaluation()?;
-
-    // sum c_star * s - d_star = 0
-    builder.try_produce_sumcheck_subpolynomial_evaluation(
-        SumcheckSubpolynomialType::ZeroSum,
-        c_star_eval * s_eval - d_star_eval,
-        2,
-    )?;
-
-    // c_star + c_fold * c_star - chi_n = 0
-    builder.try_produce_sumcheck_subpolynomial_evaluation(
-        SumcheckSubpolynomialType::Identity,
-        c_star_eval + c_fold_eval * c_star_eval - chi_n_eval,
-        2,
-    )?;
-
-    // d_star + d_fold * d_star - chi_m = 0
-    builder.try_produce_sumcheck_subpolynomial_evaluation(
-        SumcheckSubpolynomialType::Identity,
-        d_star_eval + d_fold_eval * d_star_eval - chi_m_eval,
-        2,
-    )?;
-
-    // d_fold * chi_m - d_fold = 0
-    builder.try_produce_sumcheck_subpolynomial_evaluation(
-        SumcheckSubpolynomialType::Identity,
-        d_fold_eval * (chi_m_eval - S::ONE),
-        2,
-    )?;
-
-    Ok(())
 }
 
 /// Below are the mappings between the names of the parameters in the math and the code
